Adds flag for photo segmentation - Issue #1
Adds the flag --photo-seg to the run_samseg python script. The flag
currently just disables the affine template registration as the photo
volumes are missing the cerebelum and require separate registration. As
the registration is disabled the photo-seg flag requires the filepath
of a registered template as input.

The plan is now to change the behaviour of the bias field correction
when the photo-seg flag is active. This requires changing the bias
field basis function generation and handling the downsampling so that
there is no-downsampling in the z direction.