Adds brightness handling through bias - Issue #1
Adds brightness handling through bias field correction basis functions.
The photo-seg flag now causes a change in the specification of the bias
field basis functions so that the component vectors are full vectors of
ones for the in plane directions and are vector elements of the
identity in the out of plane direction.

It remains to handle the downsampling so that the out of plane
directions aren't adversely affected.