Add regularisation for bias calculation - Issue #1
This commit adds a regularisation parameter for fitting the brightness
bassis values when used on photos.

At first a slight mistake meant that I specified too many linear basis
funcions to be composed into the bilinear basis functions for the
brightness field. Initially I didn't notice as I was getting quite good
results, however eventually the fact that there was linear dependence
between my basis functions meant that in some cases the system matrix
would become singular and the segmentation would fail.

The first solution we tried to this was to correctly specify the basis
functions. However, this method was taking an excessive amount of time
to converge. Eventually we decided to go back to the additional basis
functions but add a regularisation penalty so that the system matrix
didn't become singular.